requirements.txt
google-adk==1.2.0
Deprecated==1.2.18